abstracted frontend into different components, finished handling all comments
