+++ settings.json
@@ -1,0 +1,5 @@
+{
+    "[typescriptreact]": {
+    "editor.defaultFormatter": "esbenp.prettier-vscode"
+}
+}
