+++ src/pages/api/users/index.ts
@@ -1,5 +1,10 @@
 import { z } from "zod";
-import { createUser, editUser, getUser, editPassword } from "../../../server/db/actions/UserAction";
+import {
+  createUser,
+  editUser,
+  getUser,
+  editPassword,
+} from "../../../server/db/actions/UserAction";
 import { userSchema } from "../../../utils/types";
 import { NextApiRequest, NextApiResponse } from "next";
 import {
@@ -8,9 +13,14 @@ import {
   HTTP_INTERNAL_SERVER_ERROR,
   HTTP_METHOD_NOT_ALLOWED,
   HTTP_NOT_FOUND,
-  HTTP_UNAUTHORIZED
+  HTTP_UNAUTHORIZED,
 } from "@/utils/consts";
-import { GenericUserErrorException, UserAlreadyExistsException, UserCredentialsIncorrectException, UserDoesNotExistException } from "@/utils/exceptions";
+import {
+  GenericUserErrorException,
+  UserAlreadyExistsException,
+  UserCredentialsIncorrectException,
+  UserDoesNotExistException,
+} from "@/utils/exceptions";
 
 export const createUserSchema = userSchema
   .omit({ hashedPassword: true })
@@ -30,7 +40,7 @@ export default async function handler(
     case "GET":
       const email = String(req.query.email);
       try {
-        const user = await getUser(email); 
+        const user = await getUser(email);
         res.status(200).send({
           data: user,
         });
@@ -50,7 +60,7 @@ export default async function handler(
 
     case "PUT":
       await editUserHandler(req, res);
-      break;      
+      break;
 
     default:
       res.status(HTTP_METHOD_NOT_ALLOWED).json({
@@ -89,59 +99,57 @@ async function createUserHandler(req: NextApiRequest, res: NextApiResponse) {
 
 async function editUserHandler(req: NextApiRequest, res: NextApiResponse) {
   const { type } = req.query;
-    if (type === "info") {
-      // Editing user profile
-      try {
-        const result = await editUser(req.body);
-        res.status(200).send({
-          data: result,
+  if (type === "info") {
+    // Editing user profile
+    try {
+      const result = await editUser(req.body);
+      res.status(200).send({
+        data: result,
+      });
+      return;
+    } catch (e: unknown) {
+      if (e instanceof GenericUserErrorException) {
+        res.status(HTTP_BAD_REQUEST).json({
+          error: "User with email already exists",
         });
         return;
-      } catch (e: unknown) {
-        if (e instanceof GenericUserErrorException) {
-          res.status(HTTP_BAD_REQUEST).json({
-            error: "User with email already exists"
-          });
-          return;
-        }
-        if (e instanceof UserDoesNotExistException) {
-          res.status(HTTP_NOT_FOUND).json({
-            error: (e as Error).message,
-          });
-          return;
-        }
-        res.status(HTTP_INTERNAL_SERVER_ERROR).json({
+      }
+      if (e instanceof UserDoesNotExistException) {
+        res.status(HTTP_NOT_FOUND).json({
           error: (e as Error).message,
         });
         return;
-      } 
+      }
+      res.status(HTTP_INTERNAL_SERVER_ERROR).json({
+        error: (e as Error).message,
+      });
+      return;
     }
-
-    else if (type === "password") {
-      // Editing user password
-      try {
-        const result = await editPassword(req.body);
-        res.status(200).send({
-          data: result,
-        });
-        return;
-      } catch (e: unknown) {
-        if (e instanceof UserDoesNotExistException) {
-          res.status(HTTP_NOT_FOUND).json({
-            error: (e as Error).message,
-          });
-          return;
-        }
-        if (e instanceof UserCredentialsIncorrectException) {
-          res.status(HTTP_UNAUTHORIZED).json({
-            error: "Old password is incorrect",
-          });
-          return;
-        }
-        res.status(HTTP_INTERNAL_SERVER_ERROR).json({
+  } else if (type === "password") {
+    // Editing user password
+    try {
+      const result = await editPassword(req.body);
+      res.status(200).send({
+        data: result,
+      });
+      return;
+    } catch (e: unknown) {
+      if (e instanceof UserDoesNotExistException) {
+        res.status(HTTP_NOT_FOUND).json({
           error: (e as Error).message,
         });
         return;
-      } 
+      }
+      if (e instanceof UserCredentialsIncorrectException) {
+        res.status(HTTP_UNAUTHORIZED).json({
+          error: "Old password is incorrect",
+        });
+        return;
+      }
+      res.status(HTTP_INTERNAL_SERVER_ERROR).json({
+        error: (e as Error).message,
+      });
+      return;
     }
+  }
 }
